+++ tests/export_default_ts/function_as.ts
@@ -1,0 +1,1 @@
+export default (function log(){} as typeof console.log);
